% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data.gen}
\alias{data.gen}
\title{Generate a simulated dataset.}
\usage{
data.gen(n = 5000)
}
\arguments{
\item{n}{sample size}
}
\value{
a simulated data set with K=3 time periods
}
\description{
Generate a simulated dataset.
}
