% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DR.f.point}
\alias{DR.f.point}
\title{DR estimator (point estimation)}
\usage{
DR.f.point(
  data,
  q,
  Outcome.formula,
  PS.formula,
  S.formula,
  Var.formula = Var.formula,
  MSQM.formula,
  A.names,
  Outcome.name,
  start.value,
  tau_n
)
}
\description{
DR estimator (point estimation)
}
