% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DR.f.SE}
\alias{DR.f.SE}
\title{DR estimator (standard error)}
\usage{
DR.f.SE(
  data,
  q,
  Outcome.formula,
  PS.formula,
  S.formula,
  Var.formula = Var.formula,
  MSQM.formula,
  A.names,
  Outcome.name,
  tau_n,
  theta,
  alpha,
  alphas,
  delta,
  eta,
  Ua,
  Ia,
  Ub,
  Ib
)
}
\description{
DR estimator (standard error)
}
