% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ICR.f.point}
\alias{ICR.f.point}
\title{ICR estimator (point estimation)}
\usage{
ICR.f.point(
  data,
  q,
  Outcome.formula,
  Var.formula,
  MSQM.formula,
  A.names,
  Outcome.name,
  start.value,
  icr_coef = 0
)
}
\description{
ICR estimator (point estimation)
}
