% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ICR.f.SE}
\alias{ICR.f.SE}
\title{ICR estimator (standard error)}
\usage{
ICR.f.SE(
  data,
  q,
  Outcome.formula,
  Var.formula,
  MSQM.formula,
  A.names,
  Outcome.name,
  delta,
  eta,
  theta,
  tau_n
)
}
\description{
ICR estimator (standard error)
}
