% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{msqm.fit}
\alias{msqm.fit}
\title{Estimate the Marginal Structural Quantile Model (MSQM)}
\usage{
msqm.fit(
  data,
  q,
  PS.formula,
  S.formula,
  Outcome.formula,
  Var.formula,
  MSQM.formula,
  A.names,
  Outcome.name
)
}
\arguments{
\item{data}{a data frame}

\item{q}{desired lower quantile (ranged between 0 and 1)}

\item{PS.formula}{a list of formulas for the propensity score models in each time period}

\item{S.formula}{a list of formulas for the stabilized weights in each time period}

\item{Outcome.formula}{a list of formulas for mean structure of the outcome model in each time period}

\item{Var.formula}{a list of formulas for the variance structure of the outcome model in each time period}

\item{MSQM.formula}{Specification of the MSQM model}

\item{A.names}{the column name of treatment from period 1 to period K}

\item{Outcome.names}{the column name of the outcome (measured at the end of period K)}
}
\value{
Point, standard error, and confidence interval estimates based on the inverse
probability weighting (IPW) approach, the iterative conditional regression (ICR) approach, and the
doubly robust (DR) approach.
}
\description{
Estimate and inference on MSQM based on four approaches (IPW, ICR, and DR)
}
\examples{
#### Step 1: generate data
set.seed(12345)
data=data.gen(n=500)

head(round(data,2))
#     X11   X12   X21  X22   X31   X32 A1 A2 A3      Y
# 1 -1.21 -0.44  0.23 1.65 -0.03  0.87  1  1  0 -10.84
# 2  1.79  1.23  1.30 0.97  2.35  5.36  1  1  0  11.59
# 3  0.81  1.29  1.05 1.09  0.86  2.30  1  1  1   0.43
# 4 -0.03 -1.84  2.63 2.77 -0.08  1.87  1  1  0  -4.04
# 5 -3.88 -0.57  1.37 1.05 -0.48  0.09  1  0  0  -3.59
# 6 -1.07  1.43 -0.67 0.76  0.57 -1.08  1  0  1 -17.87
# Note: (X11, X12): baseline covariates; A1: treatment in period 1; (X21, X22): covariates at beginning of period 2;
# A2: treatment in period 2; (X31, X32): covariates at beginning of period 2; A3: treatment in period 3;
# Y: outcome measured after A3

#### Step 2: Specification
MSQM.formula=as.formula(Y~A1+A2+A3) # the MSQM model
q=0.5 # the desired quantile of interest (median here)
A.names=c("A1","A2","A3") # treatment variables
Outcome.name="Y" # outcome variable

# specification of the stabilized weights (k=1,2,3 sequentially)
S.formula=list(as.formula(A1~1),
               as.formula(A2~A1),
               as.formula(A3~A2))
# specification of the principal score model (k=1,2,3 sequentially)
PS.formula=list(as.formula(A1~X11+X12),
                as.formula(A2~A1+X21+X22),
                as.formula(A3~A2+X31+X32))
# specification of outcome mean (k=1,2,3 sequentially)
Outcome.formula=list(as.formula(Y~A1+A2+A3+X11+X12),
                     as.formula(Y~A1+A2+A3+X11+X12+X21+X22),
                     as.formula(Y~A1+A2+A3+X11+X12+X21+X22+X31+X32))
# specification of outcome variance (k=1,2,3 sequentially)
Var.formula =list(as.formula(~A1+A2+A3),
                  as.formula(~A1+A2+A3),
                  as.formula(~A1+A2+A3))

#### Step 3: fit the MSQM
obj = msqm.fit(data,q,PS.formula,S.formula,Outcome.formula,Var.formula,
         MSQM.formula,A.names,Outcome.name)
# point estimates, SE, and 95\% CI of the MSQM coefficients
obj$MSQM_coef
# the variance-covariance matrix of the MSQM coefficients
obj$MSQM_vcov
}
