% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mysummary_rq}
\alias{mysummary_rq}
\title{Returns a summary list for a quantile regression fit.}
\usage{
mysummary_rq(
  object,
  se = NULL,
  covariance = FALSE,
  hs = TRUE,
  U = NULL,
  gamma = 0.7,
  ...
)
}
\description{
This function is slightly adapted based on the summary.rq function in the quantreg package
}
